requirements.txt
streamlit
opencv-python-headless
numpy
pyyaml
setuptools
twilio

